+++ functions.py
@@ -1,0 +1,11 @@
+def get_todos(filepath='files/todos.txt'):
+	"""Function to read to-do list from file"""
+	with open(filepath, 'r') as file_local:
+		todos_local = file_local.readlines()
+	return todos_local
+
+		
+def write_todos(todos_local,filepath='files/todos.txt'):
+	"""Function to write to-do list to file"""
+	with open(filepath, 'w') as file_local:
+            file_local.writelines(todos_local)
